Only allow certain users to delete an item

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -51,10 +51,17 @@ const Mutation = {
     const where = { id: args.id };
 
     // 1. Find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
 
     // 2. Check if they own that item, or have the permissions
-    // todo
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error(`You don't have permission to do that!`);
+    }
 
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
